demo/demo/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
dateparser==0.7.6
Django==2.1.15
django-cors-headers==3.4.0
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
futures==3.1.1
gevent==20.6.2
greenlet==0.4.16
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
kombu==4.6.11
lml==0.0.9
numpy==1.19.0
openpyxl==3.0.4
pandas==1.0.5
pika==1.1.0
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
texttable==1.6.2
tzlocal==2.1
urllib3==1.25.9
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
